Make XInput/XInput2 bindings match C headers
 * Move structs and functions that are defined in XInput2 headers
   (X11/extensions/{XInput2.h, XI2.h} from xinput mod to xinput2 mod.

   Several of the XInput2 structs were duplicated in xinput and xinput2,
   so only the ones in the xinput2 mod have been kept.

 * Correct the name of the struct generated by the x11_link!() macro
   for xinput, xinput2
